+++ source/renderer/app/components/staking/DelegationInfo.js
@@ -12,23 +12,23 @@ const messages = defineMessages({
   heading: {
     id: 'paper.delegation.notification.heading',
     defaultMessage: '!!!Cardano decentralization',
-    description: 'Headline for the Decentralisation notification.',
+    description: 'Headline for the Decentralization notification.',
   },
   info: {
     id: 'paper.delegation.notification.info',
     defaultMessage:
       '!!!Cardano will soon start its transition from a federated to a decentralized system. This will mark the start of stakeholders being able to earn rewards for participating in the running of the network. They will be able to participate directly in the process of staking or can delegate their stake to stake pools to earn rewards in ada.',
-    description: 'Info for the Decentralisation notification.',
+    description: 'Info for the Decentralization notification.',
   },
   timeLeftDesc: {
     id: 'paper.delegation.notification.timeLeftDesc',
     defaultMessage: '!!!Rewards begin in',
-    description: 'Description for the Decentralisation notification.',
+    description: 'Description for the Decentralization notification.',
   },
   buttonLabel: {
     id: 'paper.delegation.notification.buttonLabel',
     defaultMessage: '!!!Learn more',
-    description: 'Button Label for the Decentralisation notification.',
+    description: 'Button Label for the Decentralization notification.',
   },
 });