+++ storybook/stories/Delegation.stories.js
@@ -1,15 +1,21 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, date } from '@storybook/addon-knobs';
+import { linkTo } from '@storybook/addon-links';
+import { withKnobs, date, number } from '@storybook/addon-knobs';
 import StoryLayout from './support/StoryLayout';
 import StoryProvider from './support/StoryProvider';
 import StoryDecorator from './support/StoryDecorator';
 import { CATEGORIES_BY_NAME } from '../../source/renderer/app/config/sidebarConfig';
+
+// Helpers
+import DelegationProgressWithNavigation from '../../source/renderer/app/components/staking/layouts/DelegationProgressWithNavigation';
+
+// Screens
 import Delegation from '../../source/renderer/app/components/staking/Delegation';
 import DelegationProgress from '../../source/renderer/app/components/staking/DelegationProgress';
 
-const percentage = 10;
+const defaultPercentage = 10;
 const defaultStartDateTime = new Date('Jun 01 2019');
 const startDateTimeKnob = (name, defaultValue) => {
   const stringTimestamp = date(name, defaultValue);
@@ -19,15 +25,38 @@ const startDateTimeKnob = (name, defaultValue) => {
 storiesOf('DelegationScreens', module)
   .addDecorator((story, context) => {
     const storyWithKnobs = withKnobs(story, context);
+    const getItemFromContext = () =>
+      context.story
+        .replace('Decentralization Progress', 'info')
+        .toLocaleLowerCase();
+
+    let activeSidebarCategory = null;
+    if (context.story === 'Decentralization Progress') {
+      activeSidebarCategory = CATEGORIES_BY_NAME.DELEGATION_PROGRESS.route;
+    } else {
+      activeSidebarCategory = CATEGORIES_BY_NAME.DELEGATION.route;
+    }
 
     return (
       <StoryDecorator>
         <StoryProvider>
           <StoryLayout
-            activeSidebarCategory={CATEGORIES_BY_NAME.DELEGATION.route}
+            activeSidebarCategory={activeSidebarCategory}
             storyName={context.story}
           >
-            {storyWithKnobs}
+            {context.story === 'Decentralization Progress' ? (
+              <DelegationProgressWithNavigation
+                isActiveScreen={item => item === getItemFromContext()}
+                onDelegationProgressNavItemClick={linkTo(
+                  'DelegationScreens',
+                  () => 'Decentralization Progress'
+                )}
+              >
+                {storyWithKnobs}
+              </DelegationProgressWithNavigation>
+            ) : (
+              storyWithKnobs
+            )}
           </StoryLayout>
         </StoryProvider>
       </StoryDecorator>
@@ -35,19 +64,22 @@ storiesOf('DelegationScreens', module)
   })
   // ====== Stories ======
 
-  .add('Start of decentralisation notification', () => (
-    <div>
-      <Delegation
-        currentLocale="en-US"
-        startDateTime={startDateTimeKnob(
-          'Delegation Start DateTime',
-          defaultStartDateTime
-        )}
-      />
-    </div>
+  .add('Start of decentralization notification', () => (
+    <Delegation
+      currentLocale="en-US"
+      startDateTime={startDateTimeKnob(
+        'Delegation Start DateTime',
+        defaultStartDateTime
+      )}
+    />
   ))
-  .add('Progress', () => (
-    <div>
-      <DelegationProgress percentage={percentage} />
-    </div>
+  .add('Decentralization Progress', () => (
+    <DelegationProgress
+      percentage={number('Percentage', defaultPercentage, {
+        min: 0,
+        max: 100,
+        step: 1,
+        range: true,
+      })}
+    />
   ));