+++ storybook/stories/support/StoryLayout.js
@@ -33,9 +33,15 @@ type Props = {
   stores?: ?{},
 };
 
-const sidebarCategories = [
+const CATEGORIES_WITH_DELEGATION = [
   CATEGORIES_BY_NAME.WALLETS,
   CATEGORIES_BY_NAME.DELEGATION,
+  CATEGORIES_BY_NAME.SETTINGS,
+];
+
+const CATEGORIES_WITH_DELEGATION_PROGRESS = [
+  CATEGORIES_BY_NAME.WALLETS,
+  CATEGORIES_BY_NAME.DELEGATION_PROGRESS,
   CATEGORIES_BY_NAME.SETTINGS,
 ];
 
@@ -70,7 +76,11 @@ export default class StoryLayout extends Component<Props> {
         }}
       >
         <SidebarLayout
-          sidebar={this.getSidebar(activeSidebarCategory, sidebarMenus)}
+          sidebar={this.getSidebar(
+            storyName,
+            activeSidebarCategory,
+            sidebarMenus
+          )}
           topbar={this.getTopbar(
             activeSidebarCategory,
             activeWallet,
@@ -85,8 +95,7 @@ export default class StoryLayout extends Component<Props> {
     );
   }
 
-  @observable
-  isShowingSubMenus =
+  @observable isShowingSubMenus =
     this.props.activeSidebarCategory === '/wallets' && !!this.props.children;
 
   getSidebarWallets = (wallets: Array<Wallet>): Array<SidebarWalletType> =>
@@ -114,20 +123,33 @@ export default class StoryLayout extends Component<Props> {
     },
   });
 
-  getSidebar = (activeSidebarCategory: string, sidebarMenus: SidebarMenus) => (
-    <Sidebar
-      categories={sidebarCategories}
-      activeSidebarCategory={activeSidebarCategory}
-      menus={sidebarMenus}
-      isShowingSubMenus={this.isShowingSubMenus}
-      onCategoryClicked={action('onCategoryClicked')}
-      isDialogOpen={() => false}
-      onAddWallet={action('onAddWallet')}
-      openDialogAction={action('openDialog')}
-      onSubmitSupportRequest={() => {}}
-      pathname="/"
-    />
-  );
+  getSidebar = (
+    storyName: string,
+    activeSidebarCategory: string,
+    sidebarMenus: SidebarMenus
+  ) => {
+    let sidebarCategories = null;
+    if (storyName === 'Decentralization Progress') {
+      sidebarCategories = CATEGORIES_WITH_DELEGATION_PROGRESS;
+    } else {
+      sidebarCategories = CATEGORIES_WITH_DELEGATION;
+    }
+
+    return (
+      <Sidebar
+        categories={sidebarCategories}
+        activeSidebarCategory={activeSidebarCategory}
+        menus={sidebarMenus}
+        isShowingSubMenus={this.isShowingSubMenus}
+        onCategoryClicked={action('onCategoryClicked')}
+        isDialogOpen={() => false}
+        onAddWallet={action('onAddWallet')}
+        openDialogAction={action('openDialog')}
+        onSubmitSupportRequest={() => {}}
+        pathname="/"
+      />
+    );
+  };
 
   getTopbar = (
     activeSidebarCategory: string,